Fixed a error causing COG to not run. General TESTING STARTS!
Testing Now available. From here on we need to adjust height and freq
values. Along with filling out the rest of the biome sets for preferered
generations. Chunks currently are pretty damn full with sparse /
sporadic and normal ore veins. Need to find a place for it all without
giving to much.